refactor: oauth_accounts 함수 모듈을 class << self로 전환 및 UserMatcher/UsernameSuggester를 Callbacks에 통합
- module_function → class << self로 변경, 외부 미노출 메서드는 private 스코프로 이동
- UserMatcher/UsernameSuggester를 Callbacks에 흡수 (match_user, suggest_username 등)
- 삭제: user_matcher.rb, username_suggester.rb 및 대응 테스트
- 테스트 3파일 → callbacks_test.rb 1개로 통합, private 메서드는 send로 호출
